chore: relocate base rpc controller
